@@ -1,6 +1,6 @@
 name: google_places_picker
 description: Flutter plugin for Google Places and Autocomplete.
-version: 2.1.0+3
+version: 3.0.2
 homepage: https://github.com/derTuca/flutter_google_places_picker
 
 dependencies:
@@ -13,8 +13,8 @@ dependencies:
 # For information on the generic Dart part of this file, see the
 # following page: https://www.dartlang.org/tools/pub/pubspec
 environment:
-  sdk: ">=2.1.0 <3.0.0"
-  flutter: ">=1.10.0 <2.0.0"
+  sdk: '>=2.12.0 <3.0.0'
+  flutter: '>=2.0.0'
 # The following section is specific to Flutter.
 flutter:
   plugin:
