@@ -4,11 +4,11 @@ import 'package:flutter/services.dart';
 import 'package:google_maps_webservice/places.dart';
 
 class Place {
-  double latitude;
-  double longitude;
-  String id;
-  String name;
-  String address;
+  late double latitude;
+  late double longitude;
+  late String id;
+  String? name;
+  String? address;
   List<AddressComponent> addressComponents;
 }
 
@@ -17,17 +17,17 @@ enum PlaceAutocompleteMode { MODE_OVERLAY, MODE_FULLSCREEN }
 enum TypeFilter { ADDRESS, CITIES, ESTABLISHMENT, GEOCODE, REGIONS }
 
 class LocationBias {
-  double northEastLat;
-  double northEastLng;
-  double southWestLat;
-  double southWestLng;
+  double northEastLat = 90.0;
+  double northEastLng = 180.0;
+  double southWestLat = 0.0;
+  double southWestLng = 0.0;
 }
 
 class LocationRestriction {
-  double northEastLat;
-  double northEastLng;
-  double southWestLat;
-  double southWestLng;
+  double northEastLat = 90.0;
+  double northEastLng = 180.0;
+  double southWestLat = 0.0;
+  double southWestLng = 0.0;
 }
 
 class PluginGooglePlacePicker {
@@ -35,11 +35,11 @@ class PluginGooglePlacePicker {
       const MethodChannel('plugin_google_place_picker');
 
   static Future<Place> showAutocomplete({
-    PlaceAutocompleteMode mode,
-    LocationBias bias,
-    LocationRestriction restriction,
-    TypeFilter typeFilter,
-    String countryCode,
+    required PlaceAutocompleteMode mode,
+    LocationBias? bias,
+    LocationRestriction? restriction,
+    TypeFilter? typeFilter,
+    String? countryCode,
   }) async {
     var argMap = {
       "mode": mode == PlaceAutocompleteMode.MODE_OVERLAY ? 71 : 72,
@@ -49,12 +49,12 @@ class PluginGooglePlacePicker {
       "country": countryCode
     };
     final Map placeMap =
-        await _channel.invokeMethod('showAutocomplete', argMap);
+        await (_channel.invokeMethod('showAutocomplete', argMap));
     return _initPlaceFromMap(placeMap);
   }
 
   static Future<void> initialize(
-      {String androidApiKey, String iosApiKey}) async {
+      {String? androidApiKey, String? iosApiKey}) async {
     await _channel.invokeMethod(
         'initialize', {"androidApiKey": androidApiKey, "iosApiKey": iosApiKey});
   }
@@ -83,7 +83,7 @@ class PluginGooglePlacePicker {
 //    }
   }
 
-  static String _convertFilterTypeToString(TypeFilter type) {
+  static String? _convertFilterTypeToString(TypeFilter? type) {
     if (type == null) {
       return null;
     }
@@ -99,31 +99,30 @@ class PluginGooglePlacePicker {
       case TypeFilter.REGIONS:
         return "regions";
     }
-    return "";
   }
 
-  static Map<String, double> _convertLocationBiasToMap(LocationBias bias) {
+  static Map<String, double>? _convertLocationBiasToMap(LocationBias? bias) {
     if (bias == null) {
       return null;
     }
     return {
-      "southWestLat": bias.southWestLat ?? 0.0,
-      "southWestLng": bias.southWestLng ?? 0.0,
-      "northEastLat": bias.northEastLat ?? 90.0,
-      "northEastLng": bias.northEastLng ?? 180.0
+      "southWestLat": bias.southWestLat,
+      "southWestLng": bias.southWestLng,
+      "northEastLat": bias.northEastLat,
+      "northEastLng": bias.northEastLng
     };
   }
 
-  static Map<String, double> _convertLocationRestrictionToMap(
-      LocationRestriction restriction) {
+  static Map<String, double>? _convertLocationRestrictionToMap(
+      LocationRestriction? restriction) {
     if (restriction == null) {
       return null;
     }
     return {
-      "southWestLat": restriction.southWestLat ?? 0.0,
-      "southWestLng": restriction.southWestLng ?? 0.0,
-      "northEastLat": restriction.northEastLat ?? 90.0,
-      "northEastLng": restriction.northEastLng ?? 180.0
+      "southWestLat": restriction.southWestLat,
+      "southWestLng": restriction.southWestLng,
+      "northEastLat": restriction.northEastLat,
+      "northEastLng": restriction.northEastLng
     };
   }
 }
